+++ app/src/main/java/com/beeswork/balance/data/network/request/ReorderPhotosRequest.kt
@@ -3,5 +3,5 @@ package com.beeswork.balance.data.network.request
 data class ReorderPhotosRequest(
     val accountId: String,
     val identityToken: String,
-    val photoOrders: Map<String, Long>
+    val photoOrders: Map<String, Int>
 )
